Set prefabs for cube tiling. Set force appropriately
Prefabs are created to be able to modify all the subset tiles in larger
aspect. Carefully. adjust forces in play and set rigidbodies to get the
right feeling. Currently, works for the small scope. Need to check for
how the impact from a great distance reacts to hitting the cubes.